backend/requirements.txt
fastapi==0.110.0
pydantic==2.6.3
uvicorn
sqlalchemy==2.0.25
python-multipart
passlib
httpx
mysql-connector-python
pymysql
pytest
pytest-asyncio
coverage
python-dotenv==1.0.1
snowflake-connector-python
pandas
openpyxl
starlette
openai
redis
